refactor: Rename JiraCustomFieldValueReplacementResolver and remove unused replacement methods
