+++ api-channel-jira/src/main/java/com/synopsys/integration/alert/api/channel/jira/distribution/JiraIssueCreationRequestCreator.java
@@ -14,7 +14,7 @@ import java.util.stream.Collectors;
 import com.synopsys.integration.alert.api.channel.jira.distribution.custom.JiraCustomFieldConfig;
 import com.synopsys.integration.alert.api.channel.jira.distribution.custom.JiraCustomFieldReplacementValues;
 import com.synopsys.integration.alert.api.channel.jira.distribution.custom.JiraCustomFieldResolver;
-import com.synopsys.integration.alert.api.channel.jira.distribution.custom.JiraCustomFieldValueReplacementUtils;
+import com.synopsys.integration.alert.api.channel.jira.distribution.custom.JiraCustomFieldValueReplacementResolver;
 import com.synopsys.integration.alert.api.channel.jira.distribution.custom.JiraResolvedCustomField;
 import com.synopsys.integration.alert.common.persistence.model.job.details.JiraJobCustomFieldModel;
 import com.synopsys.integration.jira.common.cloud.builder.IssueRequestModelFieldsBuilder;
@@ -55,7 +55,8 @@ public class JiraIssueCreationRequestCreator {
                                                            .setProject(projectId)
                                                            .setIssueType(issueType);
         for (JiraCustomFieldConfig customField : customFields) {
-            JiraCustomFieldValueReplacementUtils.injectReplacementFieldValue(customField, customFieldReplacementValues);
+            JiraCustomFieldValueReplacementResolver jiraCustomFieldValueReplacementResolver = new JiraCustomFieldValueReplacementResolver(customFieldReplacementValues);
+            jiraCustomFieldValueReplacementResolver.injectReplacementFieldValue(customField);
             JiraResolvedCustomField resolvedCustomField = jiraCustomFieldResolver.resolveCustomField(customField);
             fieldsBuilder.setValue(resolvedCustomField.getFieldId(), resolvedCustomField.getFieldValue());
         }
